requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<4.0
pandas
numpy
python-multipart
whitenoise
django-cors-headers